core/auth/requirements.txt
Django==5.0.7
ninja==1.11.1.1
django-ninja-jwt==5.3.2
psycopg2==2.9.9
requests==2.32.2
psycopg==3.2.3
django-extensions==3.2.3
Werkzeug==3.1.2
channels
twisted[http2,tls]
django-ninja==0.23.0
django-ninja-extra==0.16.0
django-environ==0.10.0
django-cors-headers==4.3.1
qrcode==7.4.2
pyotp==2.9.0
zxcvbn-python==4.4.28
gunicorn==21.2.0
whitenoise==6.6.0
